+++ package/domain/models/input_output.go
@@ -5,7 +5,7 @@ type InputOutput struct {
 	Id          int64   `gorm:"primaryKey"`
 	TaskId      int64   `gorm:"not null"`
 	Order       int     `gorm:"not null"`
-	TimeLimit   float64 `gorm:"not null"`
-	MemoryLimit float64 `gorm:"not null"`
+	TimeLimit   int64 `gorm:"not null"`
+	MemoryLimit int64 `gorm:"not null"`
 	Task        Task    `gorm:"foreignKey:TaskId; references:Id"`
 }
